Extract tests from test class
