+++ modules/bazi-calculator-manual/src/test/kotlin/net/twisterrob/astro/bazi/SexagenaryHourTestCaseTest.kt
@@ -13,7 +13,7 @@ class SexagenaryHourTestCaseTest {
 
 	@Test
 	fun `the test data has full cycles`() {
-		SexagenaryHourTestCase.CYCLES.forEach { cycle ->
+		SexagenaryHourTestCase.ALL_KNOWN_CYCLES.forEach { cycle ->
 			cycle shouldHaveSize 60
 			cycle.forEachIndexed { index, testCase ->
 				testCase.cyclicOrdinal shouldBe index + 1
@@ -23,7 +23,7 @@ class SexagenaryHourTestCaseTest {
 
 	@Test
 	fun `the test data is consistent with cycles`() {
-		SexagenaryHourTestCase.CYCLES.forEach { cycle ->
+		SexagenaryHourTestCase.ALL_KNOWN_CYCLES.forEach { cycle ->
 			cycle.forEach { testCase ->
 				val expectedPillar = BaZi.Pillar.sexagenaryCycle(testCase.cyclicOrdinal - 1)
 				val actualPillar = BaZi.Pillar(testCase.hourStem, testCase.hourBranch)
@@ -37,7 +37,7 @@ class SexagenaryHourTestCaseTest {
 
 	@Test
 	fun `the test data is consistent with dates`() {
-		SexagenaryHourTestCase.CYCLES.forEach { cycle ->
+		SexagenaryHourTestCase.ALL_KNOWN_CYCLES.forEach { cycle ->
 			cycle.forEach { testCase ->
 				testCase.startTime shouldBeBefore testCase.endTime
 				testCase.midTime shouldBeAfter testCase.startTime
